+++ R/esp_get_grid_ESDAC.R
@@ -1,16 +1,17 @@
-#' Get \pkg{sf} polygons of the national geographic grids provided by ESDAC
+#' Get \CRANpkg{sf} polygons of the national geographic grids provided by ESDAC
 #'
 #' @description
-#' Loads a \pkg{sf} polygon with the geographic grids of Spain as provided by the
-#' European Soil Data Centre (ESDAC).
+#' Loads a \CRANpkg{sf} polygon with the geographic grids of Spain as provided
+#' by the European Soil Data Centre (ESDAC).
 #'
 #' @family grids
 #'
-#' @return A \pkg{sf} polygon
+#' @return A \CRANpkg{sf} polygon
 #'
 #'
 #' @source
-#' [EEA reference grid](https://esdac.jrc.ec.europa.eu/content/european-reference-grids).
+#' [EEA reference
+#' grid](https://esdac.jrc.ec.europa.eu/content/european-reference-grids).
 #'
 #' @references
 #' - Panagos P., Van Liedekerke M., Jones A., Montanarella L., "European Soil
@@ -58,12 +59,18 @@ esp_get_grid_ESDAC <- function(resolution = 10,
 
   # Url
   if (res == 10) {
-    url <- "https://esdac.jrc.ec.europa.eu/Library/Reference_Grids/Grids/grids_for_single_eu25_countries_etrs_laea_10k.zip"
+    url <- paste0(
+      "https://esdac.jrc.ec.europa.eu/Library/Reference_Grids/",
+      "Grids/grids_for_single_eu25_countries_etrs_laea_10k.zip"
+    )
     filename <- "grids_for_single_eu25_countries_etrs_laea_10k.zip"
     init_grid <- "grid_spain_etrs_laea_10k.shp"
   } else {
     # nocov start
-    url <- "https://esdac.jrc.ec.europa.eu/Library/Reference_Grids/Grids/grid_spain_etrs_laea_1k.zip"
+    url <- paste0(
+      "https://esdac.jrc.ec.europa.eu/Library/Reference_Grids/",
+      "Grids/grid_spain_etrs_laea_1k.zip"
+    )
     filename <- "grid_spain_etrs_laea_1k.zip"
     init_grid <- "grid_spain_etrs_laea_1k.shp"
     # nocov end
@@ -81,10 +88,7 @@ esp_get_grid_ESDAC <- function(resolution = 10,
   if (update_cache || isFALSE(localfile)) {
     dwnload <- TRUE
     if (verbose) {
-      message(
-        "Downloading file from ",
-        url
-      )
+      message("Downloading file from ", url)
     }
     if (verbose && update_cache) {
       message("\nUpdating cache")
@@ -98,17 +102,17 @@ esp_get_grid_ESDAC <- function(resolution = 10,
 
   # Downloading
   if (dwnload) {
-    err_dwnload <- try(download.file(url, filepath,
-      quiet = isFALSE(verbose),
-      mode = "wb"
-    ), silent = TRUE)
+    err_dwnload <- try(
+      download.file(url, filepath, quiet = isFALSE(verbose), mode = "wb"),
+      silent = TRUE
+    )
     # nocov start
     if (inherits(err_dwnload, "try-error")) {
       if (verbose) message("Retrying query")
-      err_dwnload <- try(download.file(url, filepath,
-        quiet = isFALSE(verbose),
-        mode = "wb"
-      ), silent = TRUE)
+      err_dwnload <- try(
+        download.file(url, filepath, quiet = isFALSE(verbose), mode = "wb"),
+        silent = TRUE
+      )
     }
 
     # If not then message
@@ -135,11 +139,7 @@ esp_get_grid_ESDAC <- function(resolution = 10,
   }
 
   err_onload <- try(
-    sf::st_read(
-      init_grid,
-      quiet = isFALSE(verbose),
-      stringsAsFactors = FALSE
-    ),
+    sf::st_read(init_grid, quiet = isFALSE(verbose), stringsAsFactors = FALSE),
     silent = TRUE
   )
   # nocov start