+++ R/esp_get_gridmap.R
@@ -1,15 +1,15 @@
-#' Get a \pkg{sf} hexbin or squared polygon of Spain
+#' Get a \CRANpkg{sf} hexbin or squared polygon of Spain
 #'
 #' @description
-#' Loads a hexbin map (\pkg{sf} object) or a map of squares with the boundaries of
-#' the provinces or autonomous communities of Spain.
+#' Loads a hexbin map (\CRANpkg{sf} object) or a map of squares with the
+#' boundaries of the provinces or autonomous communities of Spain.
 #'
 #' @rdname esp_get_gridmap
 #' @name  esp_get_gridmap
 #'
 #' @family political
 #'
-#' @return A \pkg{sf} POLYGON object.
+#' @return A \CRANpkg{sf} POLYGON object.
 #'
 #'
 #' @export
@@ -61,8 +61,7 @@ esp_get_hex_prov <- function(prov = NULL) {
 
     dfcpro <- mapSpain::esp_codelist
     dfcpro <- unique(dfcpro[, c("nuts3.code", "cpro")])
-    cprocodes <-
-      unique(dfcpro[dfcpro$nuts3.code %in% region, ]$cpro)
+    cprocodes <- unique(dfcpro[dfcpro$nuts3.code %in% region, ]$cpro)
 
     data_sf <- data_sf[data_sf$cpro %in% cprocodes, ]
   }
@@ -128,8 +127,7 @@ esp_get_grid_prov <- function(prov = NULL) {
 
     dfcpro <- mapSpain::esp_codelist
     dfcpro <- unique(dfcpro[, c("nuts3.code", "cpro")])
-    cprocodes <-
-      unique(dfcpro[dfcpro$nuts3.code %in% region, ]$cpro)
+    cprocodes <- unique(dfcpro[dfcpro$nuts3.code %in% region, ]$cpro)
 
     data_sf <- data_sf[data_sf$cpro %in% cprocodes, ]
   }